make get players catch error and use arrow function to declare it
