+++ src/Score/score.js
@@ -3,7 +3,8 @@
 async function getData(url) {
   const response = await fetch(url);
   const data = await response.json();
-  return data;
+  const list  = getFilteredList(data.result)
+  return list;
 }
 
 async function postData(url, data) {
@@ -17,6 +18,20 @@ async function postData(url, data) {
 
   const feedback = await response.json();
   return feedback;
+}
+
+const getFilteredList = (arr) =>  {
+  let result = [];
+  arr.forEach((player) => {
+    const item = result.find((obj) => obj.user === player.user);
+    if (!item) {
+      result.push(player);
+    } else if (item) {
+      item.score = item.score > player.score ? item.score : player.score;
+    }
+  });
+  result = result.sort((a, b) => b.score - a.score);
+  return result;
 }
 
 export { getData, postData };
